pyproject.toml
[project]
name = "support-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.6.3",
    "flask-sqlalchemy==3.0.5",
    "jq>=1.8.0",
    "langchain>=0.3.20",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.19",
    "langchain-openai>=0.3.8",
    "load-dotenv>=0.1.0",
    "openai>=1.66.3",
    "rank-bm25>=0.2.2",
    "sqlalchemy>=1.4.54",
    "telebot>=0.0.5",
    "pymorphy2>=0.9.1",
    "pymorphy2-dicts-ru>=2.4.417127.4579844",
]
